# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/companies/{id}": {
      "put": {
        "tags": [
          "Company"
        ],
        "operationId": "updateCompany",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/validations": {
      "post": {
        "tags": [
          "Validations"
        ],
        "summary": "Submit a validation (Accept request)",
        "operationId": "submitValidation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitValidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/validation-requests/{id}/reject": {
      "post": {
        "tags": [
          "Validations"
        ],
        "summary": "Reject a validation request",
        "operationId": "rejectRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/offers": {
      "post": {
        "tags": [
          "Job Offers"
        ],
        "operationId": "createOffer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobOfferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{id}/publish": {
      "post": {
        "tags": [
          "Job Offers"
        ],
        "operationId": "publishOffer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/companies": {
      "post": {
        "tags": [
          "Company"
        ],
        "operationId": "createCompany",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/me/validation-requests": {
      "get": {
        "tags": [
          "Validations"
        ],
        "summary": "Incoming requests asking me to validate others",
        "operationId": "getIncomingRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "PENDING"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Validations"
        ],
        "summary": "Create a validation request",
        "operationId": "requestValidation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRequestPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/candidates/me/skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "summary": "List own skills with levels and scores",
        "operationId": "getMySkills",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Skills"
        ],
        "summary": "Add a skill to profile",
        "operationId": "addSkill",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSkillRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/candidates/me/matches/{offerId}/interest": {
      "post": {
        "tags": [
          "Matches (Candidate)"
        ],
        "summary": "Mark 'Me interesa' (Triggers profile reveal)",
        "operationId": "markInterest",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/candidates/me/matches/{offerId}/decline": {
      "post": {
        "tags": [
          "Matches (Candidate)"
        ],
        "summary": "Mark 'No me interesa'",
        "operationId": "declineInterest",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/skills": {
      "get": {
        "tags": [
          "Skills"
        ],
        "summary": "Search skill catalog",
        "operationId": "searchSkills",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": ""
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
    "/skills/{skillId}/suggested-validators": {
      "get": {
        "tags": [
          "Validations"
        ],
        "summary": "Get suggested validators for a specific skill",
        "operationId": "getSuggestedValidators",
        "parameters": [
          {
            "name": "skillId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuggestedValidator"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/offers/{offerId}/candidates": {
      "get": {
        "tags": [
          "Matches (Recruiter)"
        ],
        "summary": "List matched candidates for a specific offer",
        "operationId": "getMatchedCandidatesForOffer",
        "parameters": [
          {
            "name": "offerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {}
                }
              }
            }
          }
        }
      }
    },
...
    "/candidates/me/skills/{id}": {
      "delete": {
        "tags": [
          "Skills"
        ],
        "summary": "Remove a skill from profile",
        "operationId": "removeSkill",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompanyRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "cultureDescription": {
            "type": "string"
          }
        }
      },
...
      "SubmitValidationRequest": {
        "type": "object",
        "properties": {
          "requestId": {
            "type": "string",
            "format": "uuid"
          },
          "assignedLevel": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          }
        }
      },
      "JobOfferRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "modality": {
            "type": "string"
          },
          "seniority": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "salaryMin": {
            "type": "number",
            "format": "double"
          },
          "salaryMax": {
            "type": "number",
            "format": "double"
          },
          "location": {
            "type": "string"
          }
        }
      },
...
      "ValidationRequestPayload": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "validatorId": {
            "type": "string",
            "format": "uuid"
          },
          "relationType": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "AddSkillRequest": {
        "type": "object",
        "properties": {
          "skillId": {
            "type": "string",
            "format": "uuid"
          },
          "experienceRange": {
            "type": "string"
          }
        }
      },
...
      "SuggestedValidator": {
        "type": "object",
        "properties": {
          "candidateId": {
            "type": "string",
            "format": "uuid"
          },
          "fullName": {
            "type": "string"
          },
          "seniority": {
            "type": "string"
          },
          "confidenceScore": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
    }
  }
}